Assume llm output is gzipped, move gzip logic into ObjectStorage class
