+++ backend/app/services/ObjectStorage.scala
@@ -2,7 +2,7 @@ package services
 
 
 
-import java.io.InputStream
+import java.io.{ByteArrayInputStream, InputStream}
 import java.nio.file.Path
 import model.ObjectMetadata
 import software.amazon.awssdk.services.s3.model.{Delete, DeleteObjectRequest, DeleteObjectsRequest, GetObjectRequest, HeadObjectRequest, ListObjectsV2Request, ListObjectsV2Response, ObjectIdentifier, PutObjectRequest}
@@ -11,9 +11,12 @@ import software.amazon.awssdk.services.s3.presigner.model.{GetObjectPresignReque
 import utils.attempt.{Failure, IllegalStateFailure, UnknownFailure}
 import utils.aws.{AwsErrors, S3Client}
 
+import java.nio.charset.StandardCharsets
 import scala.jdk.CollectionConverters._
 import scala.util.control.NonFatal
 import java.time.Duration
+import java.util.zip.GZIPInputStream
+import scala.util.Using
 
 
 trait ObjectStorage {
@@ -21,6 +24,7 @@ trait ObjectStorage {
   def putText(key: String, text: String, mimeType: Option[String]): Either[Failure, Unit]
   def get(key: String): Either[Failure, InputStream]
   def getSignedUrl(key: String): Either[Failure, String]
+  def getGzippedText(key: String): Either[Failure, String]
   def getUploadSignedUrl(key: String): Either[Failure, String]
   def getMetadata(key: String): Either[Failure, ObjectMetadata]
   def delete(key: String): Either[Failure, Unit]
@@ -46,6 +50,22 @@ class S3ObjectStorage private(client: S3Client, presigner: S3Presigner,  bucket:
       .build()
 
     run(client.s3.getObject(getObjectRequest))
+  }
+
+  private def unzipBytes (data: Array[Byte]): String = {
+    Using.resource(new ByteArrayInputStream(data)){ byteStream =>
+      Using.resource(new GZIPInputStream(byteStream)) { gzipStream =>
+        val decompressedData = gzipStream.readAllBytes()
+        new String(decompressedData, StandardCharsets.UTF_8)
+      }
+    }
+  }
+
+  def getGzippedText(key: String): Either[Failure, String] = {
+    get(key).map {stream =>
+      val allBytes = stream.readAllBytes()
+      unzipBytes(allBytes)
+    }
   }
 
   def getSignedUrl(key: String): Either[Failure, String] = {
